requirements.txt
pyrogram==2.0.106
tgcrypto
SQLAlchemy==2.0.31
aiosqlite==0.20.0
python-dotenv==1.0.1
aiofiles==24.1.0
